Added serveral changes:fix bug, add cutoff
Fixed bug: A bug with which make ngCGH only work on first regions of every
chromosomes in a bed files. The fix however makes ngCGH does not work
with in-line regions format.

Added cutoff: Use standard deviations as a cutoff. Any log2ratio that
are within -cutoff and +cutoff deviations are removed.

Added header for output file.